+++ packages/riverpod/lib/legacy.dart
@@ -1,0 +1,7 @@
+export 'package:state_notifier/state_notifier.dart' show StateNotifier;
+
+export 'src/providers/legacy/state_controller.dart';
+export 'src/providers/legacy/state_notifier_provider.dart'
+    hide StateNotifierProviderElement, AutoDisposeStateNotifierProviderElement;
+export 'src/providers/legacy/state_provider.dart'
+    hide StateProviderElement, AutoDisposeStateProviderElement;