+++ packages/riverpod/lib/riverpod.dart
@@ -1,40 +1,5 @@
-export 'package:state_notifier/state_notifier.dart' hide Listener, LocatorMixin;
-
 // ignore: invalid_export_of_internal_element, For the sake of backward compatibility. Remove in 3.0
 export 'src/framework.dart' show ProviderElementBase;
-export 'src/framework.dart'
-    hide
-        AsyncTransition,
-        ProviderScheduler,
-        debugCanModifyProviders,
-        Vsync,
-        ValueProviderElement,
-        ValueProvider,
-        FamilyCreate,
-        AsyncSelector,
-        FamilyBase,
-        FamilyOverrideImpl,
-        AutoDisposeProviderElementMixin,
-        FamilyOverride,
-        NotifierFamilyBase,
-        SetupFamilyOverride,
-        SetupOverride,
-        AutoDisposeNotifierFamilyBase,
-        ProviderOverride,
-        AutoDisposeFamilyBase,
-        AlwaysAliveAsyncSelector,
-        DebugGetCreateSourceHash,
-        ProviderNotifierCreate,
-        ProviderCreate,
-        computeAllTransitiveDependencies,
-        Create,
-        Node,
-        ProviderElementProxy,
-        OnError,
-        ProviderListenableOrScope,
-        AnyX,
-        ObsX,
-        ProviderElementBase;
 // ignore: invalid_export_of_internal_element, For the sake of backward compatibility. Remove in 3.0
 export 'src/providers/async_notifier.dart'
     show
@@ -43,68 +8,27 @@ export 'src/providers/async_notifier.dart'
         StreamNotifierProviderElement,
         AutoDisposeAsyncNotifierProviderElement,
         AutoDisposeStreamNotifierProviderElement;
-export 'src/providers/async_notifier.dart'
-    hide
-        AsyncNotifierProviderImpl,
-        AutoDisposeAsyncNotifierProviderImpl,
-        AutoDisposeFamilyAsyncNotifierProviderImpl,
-        FamilyAsyncNotifierProviderImpl,
-        AsyncNotifierBase,
-        AsyncNotifierProviderBase,
-        CancelAsyncSubscription,
-        BuildlessAsyncNotifier,
-        BuildlessAutoDisposeAsyncNotifier,
-        FutureHandlerProviderElementMixin,
-        FamilyStreamNotifierProviderImpl,
-        StreamNotifierProviderImpl,
-        AutoDisposeStreamNotifierProviderImpl,
-        AutoDisposeFamilyStreamNotifierProviderImpl,
-        StreamNotifierProviderBase,
-        BuildlessAutoDisposeStreamNotifier,
-        BuildlessStreamNotifier,
-        AsyncNotifierProviderElement,
-        AsyncNotifierProviderElementBase,
-        StreamNotifierProviderElement,
-        AutoDisposeAsyncNotifierProviderElement,
-        AutoDisposeStreamNotifierProviderElement;
 // ignore: invalid_export_of_internal_element, For the sake of backward compatibility. Remove in 3.0
 export 'src/providers/future_provider.dart'
     show FutureProviderElement, AutoDisposeFutureProviderElement;
-export 'src/providers/future_provider.dart'
-    hide FutureProviderElement, AutoDisposeFutureProviderElement;
-export 'src/providers/legacy/state_controller.dart';
 // ignore: invalid_export_of_internal_element, For the sake of backward compatibility. Remove in 3.0
 export 'src/providers/legacy/state_notifier_provider.dart'
     show StateNotifierProviderElement, AutoDisposeStateNotifierProviderElement;
-export 'src/providers/legacy/state_notifier_provider.dart'
-    hide StateNotifierProviderElement, AutoDisposeStateNotifierProviderElement;
 // ignore: invalid_export_of_internal_element, For the sake of backward compatibility. Remove in 3.0
 export 'src/providers/legacy/state_provider.dart'
     show StateProviderElement, AutoDisposeStateProviderElement;
-export 'src/providers/legacy/state_provider.dart'
-    hide StateProviderElement, AutoDisposeStateProviderElement;
 // ignore: invalid_export_of_internal_element, For the sake of backward compatibility. Remove in 3.0
 export 'src/providers/notifier.dart'
     show NotifierProviderElement, AutoDisposeNotifierProviderElement;
-export 'src/providers/notifier.dart'
-    hide
-        NotifierBase,
-        NotifierProviderBase,
-        AutoDisposeFamilyNotifierProviderImpl,
-        AutoDisposeNotifierProviderImpl,
-        FamilyNotifierProviderImpl,
-        NotifierProviderImpl,
-        BuildlessAutoDisposeNotifier,
-        BuildlessNotifier,
-        NotifierProviderElement,
-        AutoDisposeNotifierProviderElement;
 // ignore: invalid_export_of_internal_element, For the sake of backward compatibility. Remove in 3.0
 export 'src/providers/provider.dart'
     show ProviderElement, AutoDisposeProviderElement;
-export 'src/providers/provider.dart'
-    hide InternalProvider, ProviderElement, AutoDisposeProviderElement;
 // ignore: invalid_export_of_internal_element, For the sake of backward compatibility. Remove in 3.0
 export 'src/providers/stream_provider.dart'
     show StreamProviderElement, AutoDisposeStreamProviderElement;
-export 'src/providers/stream_provider.dart'
-    hide StreamProviderElement, AutoDisposeStreamProviderElement;
+
+@Deprecated(
+  'This is old API. Use `package:riverpod/legacy.dart` if you want to keep using it.',
+)
+export 'src/riverpod_legacy.dart';
+export 'src/riverpod_without_legacy.dart';