+++ packages/riverpod/lib/src/riverpod_without_legacy.dart
@@ -1,0 +1,79 @@
+export 'package:state_notifier/state_notifier.dart' hide Listener, LocatorMixin;
+
+export 'framework.dart'
+    hide
+        AsyncTransition,
+        ProviderScheduler,
+        debugCanModifyProviders,
+        Vsync,
+        ValueProviderElement,
+        ValueProvider,
+        FamilyCreate,
+        AsyncSelector,
+        FamilyBase,
+        FamilyOverrideImpl,
+        AutoDisposeProviderElementMixin,
+        FamilyOverride,
+        NotifierFamilyBase,
+        SetupFamilyOverride,
+        SetupOverride,
+        AutoDisposeNotifierFamilyBase,
+        ProviderOverride,
+        AutoDisposeFamilyBase,
+        AlwaysAliveAsyncSelector,
+        DebugGetCreateSourceHash,
+        ProviderNotifierCreate,
+        ProviderCreate,
+        computeAllTransitiveDependencies,
+        Create,
+        Node,
+        ProviderElementProxy,
+        OnError,
+        ProviderListenableOrScope,
+        AnyX,
+        ObsX,
+        ProviderElementBase;
+
+export 'providers/async_notifier.dart'
+    hide
+        AsyncNotifierProviderImpl,
+        AutoDisposeAsyncNotifierProviderImpl,
+        AutoDisposeFamilyAsyncNotifierProviderImpl,
+        FamilyAsyncNotifierProviderImpl,
+        AsyncNotifierBase,
+        AsyncNotifierProviderBase,
+        CancelAsyncSubscription,
+        BuildlessAsyncNotifier,
+        BuildlessAutoDisposeAsyncNotifier,
+        FutureHandlerProviderElementMixin,
+        FamilyStreamNotifierProviderImpl,
+        StreamNotifierProviderImpl,
+        AutoDisposeStreamNotifierProviderImpl,
+        AutoDisposeFamilyStreamNotifierProviderImpl,
+        StreamNotifierProviderBase,
+        BuildlessAutoDisposeStreamNotifier,
+        BuildlessStreamNotifier,
+        AsyncNotifierProviderElement,
+        AsyncNotifierProviderElementBase,
+        StreamNotifierProviderElement,
+        AutoDisposeAsyncNotifierProviderElement,
+        AutoDisposeStreamNotifierProviderElement;
+export 'providers/future_provider.dart'
+    hide FutureProviderElement, AutoDisposeFutureProviderElement;
+
+export 'providers/notifier.dart'
+    hide
+        NotifierBase,
+        NotifierProviderBase,
+        AutoDisposeFamilyNotifierProviderImpl,
+        AutoDisposeNotifierProviderImpl,
+        FamilyNotifierProviderImpl,
+        NotifierProviderImpl,
+        BuildlessAutoDisposeNotifier,
+        BuildlessNotifier,
+        NotifierProviderElement,
+        AutoDisposeNotifierProviderElement;
+export 'providers/provider.dart'
+    hide InternalProvider, ProviderElement, AutoDisposeProviderElement;
+export 'providers/stream_provider.dart'
+    hide StreamProviderElement, AutoDisposeStreamProviderElement;