+++ packages/riverpod/test/old/framework/provider_container_test.dart
@@ -1,5 +1,5 @@
 import 'package:mockito/mockito.dart';
-import 'package:riverpod/riverpod.dart';
+import 'package:riverpod/legacy.dart';
 import 'package:riverpod/src/internals.dart';
 import 'package:test/test.dart';