+++ packages/riverpod/test/src/providers/async_notifier_test.dart
@@ -5,6 +5,7 @@ import 'dart:async';
 import 'package:meta/meta.dart';
 import 'package:mockito/mockito.dart';
 import 'package:riverpod/riverpod.dart' hide ErrorListener;
+import 'package:riverpod/legacy.dart';
 import 'package:test/test.dart';
 
 import '../../utils.dart';